to do: merged tag 1.0.1 onto the trunk manually (should be functionally in sync for all of philips stims) didn't bother with compile which works on the trunk specific changes made based on no-standard improvements: -2623: ifFeature protocol settings date, calcStim temporal details and blocking error check,  getDefaultParameters gets may and nov DONE -2480: getDefaultParameters (tag:  2_6 for blocking,  default.blockingExperiments=[], ifFeature gets 2_6 in error check inputs   DONE -2386: got 2_3 reduced into getDefaultParameters DONE -2367: setReinforcementParam returns updated r DONE whole sale overwriting of setup files or files confirmed not to change on the trunk: -checkBlocking DONE -setBlockedDetials  DONE -setValuesFromMiniDatabase  DONE -selectStimulusParameters DONE -setUpPMM DONE -setShapingPMM  DONE -makeMiniDatabase DONE -getBasicFacts DONE
git-svn-id: svn://132.239.158.177/projects/ratrix/trunk@2708 d291aab3-510b-4049-a6d1-fe65b8d52d70

+++ classes/@ratrix/getBasicFacts.m
@@ -33,7 +33,11 @@ for i=1:size(subjectIDs,2)
             
         %calculate the reward the rat would get after 100 corrects
         [trialRecords(1:100).correct]=deal(true);
-        [rm rewardSizeULorMS requestRewardSizeULorMS msPenalty] = calcReinforcement(rm,trialRecords, s);
+        if ~isa(rm,'cuedReinforcement')  % cued breaks b/c it requires calcstim's details
+            [rm rewardSizeULorMS requestRewardSizeULorMS msPenalty] = calcReinforcement(rm,trialRecords, s);
+        else
+            rewardSizeULorMS=-1;
+        end
         stimName=class(sm);
         rewardName=class(rm);
         protocolName=getName(p);
@@ -61,10 +65,14 @@ for i=1:size(subjectIDs,2)
         try
             if isa(sm,'stimManager')
                 out{count,9}=getCurrentShapedValue(sm);
+                out{count,10}=getPercentCorrectionTrials(sm); %all SM need to report this if they have it!
             else
                 out{count,9}=nan;
+                out{count,10}=nan;
             end
-        catch
+        catch ex
+            warning('probably the stim manager does not have that method')
+            getReport(ex)
             keyboard
         end
 
@@ -77,6 +85,7 @@ for i=1:size(subjectIDs,2)
         out{count,7}=nan;
         out{count,8}=nan;
         out{count,9}=nan;
+        out{count,10}=nan;
     end
